+++ sumologic/sumologic_sources.go
@@ -335,7 +335,7 @@ func (s *Client) DestroySource(sourceID int, collectorID int) error {
 	return err
 }
 
-func (s *Client) GetSourceName(collectorID int, sourceName string) (*Source, error) {
+func (s *Client) GetSourceName(collectorID int64, sourceName string) (*Source, error) {
 
 	data, _, err := s.Get(
 		fmt.Sprintf("v1/collectors/%d/sources", collectorID),
